pyproject.toml
[project]
name = "trusted-router"
version = "0.1.0"
description = "TrustedRouter public alpha control plane and OpenRouter-compatible API."
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
  "boto3>=1.35.0",
  "cryptography>=43.0.0",
  "email-validator>=2.0.0",
  "eth-account>=0.13.7",
  "fastapi>=0.115.0",
  "google-cloud-bigtable>=2.26.0",
  "google-cloud-spanner>=3.50.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.20",
  "sentry-sdk[fastapi]>=2.20.0",
  "stripe>=11.0.0",
  "uvicorn[standard]>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trusted_router"]

[dependency-groups]
dev = [
  "hypothesis>=6.152.4",
  "mypy>=1.13",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=7.1.0",
  "pytest-httpx>=0.36.2",
  "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
branch = true
source = ["trusted_router"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.12"
files = ["src/trusted_router"]
follow_imports = "silent"
ignore_missing_imports = true
allow_untyped_globals = true
check_untyped_defs = true
warn_unused_ignores = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]
ignore = [
  "S101", # tests use assert
  "B008", # FastAPI Depends is intentionally used in route defaults
  "E501", # route matrices and response shapes are clearer left unwrapped in places
]
